requirements.txt
flask
pandas
openpyxl
discord.py
